requirements.txt
colorama==0.4.1
cassandra_driver==3.20.2
